Split RuleDefinitionListBase and RuleDefinitionListIndexedBase from rule_base.py
