+++ rct229/rule_engine/rule_base.py
@@ -4,7 +4,6 @@ from rct229.rule_engine.user_baseline_proposed_vals import UserBaselineProposedV
 from rct229.utils.assertions import MissingKeyException, RCTFailureException
 from rct229.utils.json_utils import slash_prefix_guarantee
 from rct229.utils.jsonpath_utils import find_all
-from rct229.utils.match_lists import match_lists
 
 
 class RuleDefinitionBase:
@@ -606,329 +605,3 @@ class RuleDefinitionBase:
         """
 
         return self.fail_msg
-
-
-class RuleDefinitionListBase(RuleDefinitionBase):
-    """
-    Baseclass for Rule Definitions that apply to each element in a list context.
-    """
-
-    def __init__(
-        self,
-        rmrs_used,
-        each_rule,
-        id=None,
-        description=None,
-        rmr_context="",
-        required_fields=None,
-        manual_check_required_msg="Manual Check Required",
-        not_applicable_msg="Not Applicable",
-    ):
-        self.each_rule = each_rule
-        super(RuleDefinitionListBase, self).__init__(
-            rmrs_used=rmrs_used,
-            id=id,
-            description=description,
-            rmr_context=rmr_context,
-            required_fields=required_fields,
-            manual_check_required_msg=manual_check_required_msg,
-            not_applicable_msg=not_applicable_msg,
-        )
-
-    def create_context_list(self, context, data=None):
-        """Generates a list of context trios
-
-        For a list-type rule, we need to create a list of contexts to pass on
-        to the sub-rule. Often, we need to match up the entries in the
-        RMR lists by name or id and then apply the sub-rule to each trio of entries.
-        This method is responsible for creating the list of trios.
-
-        This method must be overridden.
-
-        Parameters
-        ----------
-        context : UserBaselineProposedVals
-            Object containing the contexts for the user, baseline, and proposed RMRs
-        data : An optional data object.
-
-        Returns
-        -------
-        list of UserBaselineProposedVals
-            A list of context trios
-        """
-        raise NotImplementedError
-
-    def create_data(self, context, data=None):
-        """Create the data object to be passed to each_rule
-
-        This is typically overridden to collect data available at this rule's
-        level that is needed by each_rule.
-
-        This default implementation simply returns the data that was passed in.
-
-        Parameters
-        ----------
-        context : UserBaselineProposedVals
-            Object containing the user, baseline, and proposed contexts
-        data : any
-            The data object that was passed into the rule.
-
-        Returns
-        -------
-        UserBaselineProposedVals
-            Object containing the contexts for the user, baseline, and proposed
-            RMRs; an RMR's context is set to None if the corresponding flag
-            in self.rmrs_used is not set
-        """
-        return data
-
-    def list_filter(self, context_item, data=None):
-        """Function used to filter the context_list
-
-        The default implementation simply passes each list_item through (no filtering)
-
-        An inheriting rule can override this function to reduce the context list that
-        is returned from create_context_list.
-
-        NOTE: when overriding this function, it is important to know that rmd scope for none-index rmd, could be NONE.
-        It is recommended to add a NONE check for those rmd scope.
-
-        Parameters
-        ----------
-        context_item : dict
-            An item from the context_list
-        data : object
-            The data object for the rule. Note: list_filter will be used after
-            create_data is called, so this function will have access to any data
-            this rule added to the data object.
-            This implementation ignores the data argument, but overriding functions
-            are free to make use of it.
-
-        Returns
-        -------
-        list of UserBaselineProposedVals
-            A filtered list of context trios
-        """
-        return context_item
-
-    def rule_check(self, context, calc_vals=None, data=None):
-        """Overrides the base implementation to apply a rule to each entry in
-        a list
-
-        This should not be overridden. Override create_context_list() instead.
-
-        Parameters
-        ----------
-        context : UserBaselineProposedVals
-            Object containing the contexts for the user, baseline, and proposed RMRs
-        data : An optional data object. It is ignored by this base implementation.
-
-        Returns
-        -------
-        list
-            A list of rule outcomes. The each outcome in the list is augmented
-            with a name field that is the name of the entry in the context list.
-        """
-        # Create the data to be passed to each_rule
-        data = self.create_data(context, data)
-        context_list = self.create_context_list(context, data)
-        filtered_context_list = [
-            context_item
-            for context_item in context_list
-            if self.list_filter(context_item, data)
-        ]
-        outcomes = []
-
-        for ubp in filtered_context_list:
-            item_outcome = self.each_rule.evaluate(ubp, data)
-
-            # Set the id for item_outcome
-            if ubp.user and ubp.user["id"]:
-                item_outcome["id"] = ubp.user["id"]
-            elif ubp.baseline and ubp.baseline["id"]:
-                item_outcome["id"] = ubp.baseline["id"]
-            elif ubp.proposed and ubp.proposed["id"]:
-                item_outcome["id"] = ubp.proposed["id"]
-
-            outcomes.append(item_outcome)
-        return outcomes
-
-
-class RuleDefinitionListIndexedBase(RuleDefinitionListBase):
-    """
-    Baseclass for List-type Rule Definitions that use one of the RMR lists as an index list
-
-    Applicable rules typically have the form "for each ___ in the ??? RMR, ...".
-
-    Parameters
-    ----------
-    rmrs_used : UserBaselineProposedVals
-        A trio of boolen values indicating which RMRs are required by the
-        rule
-    each_rule : RuleDefinitionBase | RuleDefinitionListBase
-        The rule to be applied to each element in the list
-    index_rmr : "user" | "baseline" | "proposed"
-        Indicates the RMR to be indexed over
-    id : string
-        Unique id for the rule
-        Usually unspecified for nested rules
-    description : string
-        Rule description
-        Usually unspecified for nested rules
-    rmr_context : string
-        A json pointer into each RMR, or RMR fragment, provided to the rule.
-        For better human readability, the leading "/" may be ommitted.
-    list_path : string
-        A json path string into each RMR fragment that was produced by applying
-        rmr_context. The resulting sub-RMR fragments should be the lists to be
-        looped over. The default is "[*]" which assumes that the rmr_context is
-        the list to be looped over.
-        Note: the create_context_list() method can be overridden and
-        ignore list_context.
-        For better human readability, the leading "/" may be ommitted.
-        function (preceding _) inside the enclosing rule definition.
-    match_by : string
-        A json pointer into each element of the list, generally to a field
-        of the list element. The default is "/id" since the id is assumed to
-        be unique to the entire RMR.
-    """
-
-    def __init__(
-        self,
-        rmrs_used,
-        each_rule,
-        index_rmr,
-        id=None,
-        description=None,
-        rmr_context="",
-        list_path="[*]",
-        match_by="id",
-        required_fields=None,
-        manual_check_required_msg="Manual Check Required",
-        not_applicable_msg="Not Applicable",
-    ):
-        self.index_rmr = index_rmr
-        self.list_path = list_path
-        self.match_by = slash_prefix_guarantee(match_by)
-        super(RuleDefinitionListIndexedBase, self).__init__(
-            rmrs_used=rmrs_used,
-            each_rule=each_rule,
-            id=id,
-            description=description,
-            rmr_context=rmr_context,
-            required_fields=required_fields,
-            manual_check_required_msg=manual_check_required_msg,
-            not_applicable_msg=not_applicable_msg,
-        )
-
-    def create_context_list(self, context, data=None):
-        """Generates a list of context trios
-
-        Overrides the base implementation to create a list that has an entry
-        for each item in the index_rmr RMR, the other RMR entries are padded with
-        None for non-matches.
-
-        The resulting list can also be filtered using the list_filter field.
-
-        This may be overridden to produce lists that do not directly appear in
-        the RMR.
-
-        Parameters
-        ----------
-        context : UserBaselineProposedVals
-            Object containing the contexts for the user, baseline, and proposed RMRs.
-            The base implementation here takes the list context from the rmr context
-            and assumes each part is a list.
-        data : An optional data object. It is ignored by this base implementation.
-
-        Returns
-        -------
-        list of UserBaselineProposedVals
-            A list of context trios
-        """
-        UNKNOWN_INDEX_RMR = "Unknown index_rmr"
-        UNUSED_INDEX_RMR_MSG = "index_rmr is not being used"
-        CONTEXT_NOT_LIST = "The list contexts must be lists"
-
-        index_rmr = self.index_rmr
-        rmrs_used = self.rmrs_used
-        match_by = self.match_by
-
-        # The index RMR must be either user, baseline, or proposed
-        if index_rmr not in ["user", "baseline", "proposed"]:
-            raise ValueError(UNKNOWN_INDEX_RMR)
-
-        # The index RMR must be used
-        if (
-            (index_rmr == "user" and not self.rmrs_used.user)
-            or (index_rmr == "baseline" and not self.rmrs_used.baseline)
-            or (index_rmr == "proposed" and not self.rmrs_used.proposed)
-        ):
-            raise ValueError(CONTEXT_NOT_LIST)
-
-        # Get the list contexts
-        list_trio = UserBaselineProposedVals(
-            find_all(self.list_path, context.user) if rmrs_used.user else None,
-            find_all(self.list_path, context.baseline) if rmrs_used.baseline else None,
-            find_all(self.list_path, context.proposed) if rmrs_used.proposed else None,
-        )
-
-        # # This implementation assumes the used lists contexts are in fact lists
-        # if (
-        #     (rmrs_used.user and not isinstance(list_context_trio.user, list))
-        #     or (rmrs_used.baseline and not isinstance(list_context_trio.baseline, list))
-        #     or (rmrs_used.proposed and not isinstance(list_context_trio.proposed, list))
-        # ):
-        #     raise ValueError(CONTEXT_NOT_LIST)
-
-        user_list = None
-        baseline_list = None
-        proposed_list = None
-
-        # User indexed
-        if index_rmr == "user":
-            user_list = list_trio.user
-            context_list_len = len(user_list)
-            if rmrs_used.baseline:
-                baseline_list = match_lists(
-                    list_trio.user, list_trio.baseline, match_by
-                )
-            if rmrs_used.proposed:
-                proposed_list = match_lists(
-                    list_trio.user, list_trio.proposed, match_by
-                )
-
-        # Baseline indexed
-        elif index_rmr == "baseline":
-            baseline_list = list_trio.baseline
-            context_list_len = len(baseline_list)
-            if rmrs_used.user:
-                user_list = match_lists(list_trio.baseline, list_trio.user, match_by)
-            if rmrs_used.proposed:
-                proposed_list = match_lists(
-                    list_trio.baseline, list_trio.proposed, match_by
-                )
-
-        # Proposed indexed
-        elif index_rmr == "proposed":
-            proposed_list = list_trio.proposed
-            context_list_len = len(proposed_list)
-            if rmrs_used.user:
-                user_list = match_lists(list_trio.proposed, list_trio.user, match_by)
-            if rmrs_used.baseline:
-                baseline_list = match_lists(
-                    list_trio.proposed, list_trio.baseline, match_by
-                )
-
-        # Generate the context list
-        context_list = []
-        for index in range(context_list_len):
-            user_entry = None if user_list is None else user_list[index]
-            baseline_entry = None if baseline_list is None else baseline_list[index]
-            proposed_entry = None if proposed_list is None else proposed_list[index]
-
-            context_list.append(
-                UserBaselineProposedVals(user_entry, baseline_entry, proposed_entry)
-            )
-
-        return context_list
